Make start scripts work with symlinks
This fixes running the workbench using symlinks. I have symlinks to the
workbench start scripts in my $HOME/bin directory. Running with a
symlink from that folder would cause this error, the reason being that
the way the SCRIPT_PATH variable was created did not follow symlinks:

```
Error: Could not find or load main class workbench.WbStarter
Caused by: java.lang.ClassNotFoundException: workbench.WbStarter
```

Along with this change, the shebang is updated to look up Bash from the
environment. That's important on a Mac when you don't want to use the
ancient default Bash (3.2) that comes with the MacOS but a newer one
that's e. g. installed via Homebrew and can be found in
/usr/local/bin/bash.